ui remaster and fix: schemas

+++ apps/backend/src/auth/user.service.ts
@@ -1,4 +1,10 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import {
+  ConflictException,
+  Inject,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { User } from "./user.entity";
@@ -6,6 +12,7 @@ import { ObjectId } from "mongodb";
 import { toObjectId } from "../utils/convertion.type";
 import { UserUpdateInputDTO } from "./dto/user-update-input.dto";
 import { CloudinaryService } from "../upload/cloudinary.service";
+import { ConfigService } from "@nestjs/config";
 
 @Injectable()
 export class UserService {
@@ -13,6 +20,7 @@ export class UserService {
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
     private readonly cloudinaryService: CloudinaryService,
+    @Inject(ConfigService) configService: ConfigService
   ) {}
 
   async findByID(id: number): Promise<User | undefined> {
@@ -36,8 +44,7 @@ export class UserService {
     id: number,
     userUpdateInputDTO: UserUpdateInputDTO
   ): Promise<User> {
-
-    const { profileFile}  = userUpdateInputDTO;
+    const { profileFile } = userUpdateInputDTO;
     const user = await this.userRepository.findOne({ where: { id } });
     if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
@@ -53,12 +60,34 @@ export class UserService {
         uniqueFilename,
         "users"
       );
-      if(user.picture)
-      await this.cloudinaryService.deleteFile(user.picture);
+      if (user.picture) await this.cloudinaryService.deleteFile(user.picture);
 
       console.log(uploadResult.public_id, uploadResult.secure_url);
       user.picture = uploadResult.secure_url;
     }
     return this.userRepository.save(user);
+  }
+  async changeUserRoleToAdmin(id: number,secretKey:string): Promise<boolean> {
+    const envSecretKey = process.env.ADMIN_SECRET_KEY;
+    console.log(envSecretKey);
+    if (secretKey !== envSecretKey) {
+      throw new UnauthorizedException("Invalid secret key.");
+    }
+
+    const existingAdmin = await this.userRepository.findOne({
+      where: { role: "admin" },
+    });
+    if (existingAdmin) {
+      throw new ConflictException("Admin already exists.");
+    }
+
+    const user = await this.userRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
+    user.role = "admin";
+    await this.userRepository.save(user);
+    return true;
   }
 }
